linker/linker.c: COPYABLE_FUNC: manually specify section attributes for both the function and adj_data
it seems that on aarch64 gcc `as` doesn't treat lack of section attributes as
a cue to just use whatever the other variables in the section are using, and
so errors out even with empty section attributes.

we can't just remove the section attributes for _adj_data as well because then
the section will be seen as a debug section and the runtime linker won't
be able to find the function when we want to copy it.